requirements.txt
antlr-ast==0.8.1
antlr4-python3-runtime==4.7.2
astroid==2.4.2
click==7.1.2
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.5.3
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
